# internal/api/handler/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.createLibraryRequest:
    properties:
      media_type:
        type: string
      path:
        type: string
    type: object
...
  models.Library:
    properties:
      created_at:
        type: string
      id:
        type: string
      media_type:
        $ref: '#/definitions/models.MediaType'
      path:
        type: string
      updated_at:
        type: string
    type: object
...
  models.MediaType:
    enum:
    - movie
    - tvshow
    - episode
    - music
    type: string
    x-enum-varnames:
    - MediaTypeMovie
    - MediaTypeTVShow
    - MediaTypeEpisode
    - MediaTypeMusic
...
host: localhost:8080
info:
  contact: {}
  description: REST & WebSocket API specification for Prism, a self-hosted media server.
  title: Prism Media Server API
  version: "1.0"
paths:
...
  /libraries:
    get:
      description: Retrieve a list of all libraries (movie, tvshow, etc.) indexed
        by the server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Library'
            type: array
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Libraries
      tags:
      - Libraries
    post:
      consumes:
      - application/json
      description: Add a new media library pointing to a directory path.
      parameters:
      - description: Library creation properties
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createLibraryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Library'
        "400":
          description: Invalid request body or missing fields
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (requires Admin status)
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Library path already registered
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Library (Admin Only)
      tags:
      - Libraries
  /libraries/{id}:
    delete:
      parameters:
      - description: Library ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Library deleted successfully
        "400":
          description: Invalid library ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (requires Admin status)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Library not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Library (Admin Only)
      tags:
      - Libraries
    get:
      parameters:
      - description: Library ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Library'
        "400":
          description: Invalid library ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Library not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Library Details
      tags:
      - Libraries
  /libraries/{id}:scan:
    post:
      description: Triggers a directory scan to index new/updated files.
      parameters:
      - description: Library ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Library scan triggered
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid library ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (requires Admin status)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Library not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan Library (Admin Only)
      tags:
      - Libraries
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
  WorkerAuth:
    description: The secret API Key assigned to the transcode worker.
    in: header
    name: X-Worker-API-Key
    type: apiKey
swagger: "2.0"